[ANDROSDK-2003] Move systemInfo fields and service to network
