+++ core/src/main/java/org/hisp/dhis/android/network/systeminfo/SystemInfoService.kt
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2004-2023, University of Oslo
+ *  Copyright (c) 2004-2025, University of Oslo
  *  All rights reserved.
  *
  *  Redistribution and use in source and binary forms, with or without
@@ -25,16 +25,16 @@
  *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.android.core.systeminfo.internal
+package org.hisp.dhis.android.network.systeminfo
 
-import org.hisp.dhis.android.core.arch.api.HttpServiceClient
 import org.hisp.dhis.android.core.systeminfo.SystemInfo
+import org.hisp.dhis.android.network.common.HttpServiceClientKotlinx
 import org.hisp.dhis.android.network.common.fields.Fields
 import org.koin.core.annotation.Singleton
 
 @Singleton
-internal class SystemInfoService(private val client: HttpServiceClient) {
-    suspend fun getSystemInfo(fields: Fields<SystemInfo>): SystemInfo {
+internal class SystemInfoService(private val client: HttpServiceClientKotlinx) {
+    suspend fun getSystemInfo(fields: Fields<SystemInfo>): SystemInfoDTO {
         return client.get {
             url("system/info")
             parameters {
